Much better UCI support, including nodes, searchmoves, UCI options for Hash (and clearing), Threads, lowerbound/upperbound for aspiration windows, seldepth, currmove/currmovenumber, quiescence PV, better time management

+++ hash.h
@@ -43,7 +43,7 @@ namespace tt {
 
     class hash_t {
     public:
-        explicit hash_t(unsigned int size);
+        explicit hash_t(size_t size);
 
         ~hash_t();
 
@@ -53,7 +53,7 @@ namespace tt {
     private:
         const int BUCKET_SIZE = 4;
 
-        int num_entries;
+        size_t num_entries;
         entry_t *table;
     };
 }
